Update lmo-tippsavewertgesamt.php
Beim aktualisieren der Gesamtwertung des Tippspiel durch den Admin kam die Fehlermeldung:
PHP Warning:  A non-numeric value encountered in ../lmo/addon/tipp/lmo-tippsavewertgesamt.php on line 128

+++ lmo/addon/tipp/lmo-tippsavewertgesamt.php
@@ -125,7 +125,7 @@ for($k = 0; $k < $anzligen; $k++) {
       $dum = explode('|', $tippdaten[$i-1]);
       $op1 = $dum[0];
       // Nick
-      $op3 = substr($dum[1], 2)-1;
+      $op3 = (int)substr($dum[1], 2)-1;
       // Spieltagsnummer
       $op4 = substr($dum[1], 0, 2);
       // PP
